🔧 Implement Priority 1 & 2 fixes from combined CCR
Based on CC, Warp, and Gemini Critical Code Reviews:

Priority 1 Fixes (Critical):
✅ KG embedding metadata stamping - Now saves provider/model/dim/embedded_at
✅ HTTP timeout for Grok calls - Added 20s timeout to prevent hangs
✅ Content size validation - 100KB limit on think tool content
✅ Documentation updates - README and startup logs match actual tools

Priority 2 Fixes (Performance):
✅ Database index - Added idx_thoughts_embedding_dim for query optimization
✅ Code deduplication - Extracted cosine_similarity to utils/math.rs

All changes tested:
- cargo check ✅
- cargo clippy -- -D warnings ✅
- cargo test ✅

Total implementation time: ~45 minutes
No heroes were harmed in the making of these fixes.
