+++ src/tools/search_thoughts.rs
@@ -139,25 +139,10 @@ impl SurrealMindServer {
             q_emb.len()
         );
 
-        // Compute cosine similarity locally
-        fn cosine(a: &[f32], b: &[f32]) -> f32 {
-            if a.is_empty() || b.is_empty() || a.len() != b.len() {
-                return 0.0;
-            }
-            let mut dot = 0.0f32;
-            let mut na = 0.0f32;
-            let mut nb = 0.0f32;
-            for i in 0..a.len() {
-                dot += a[i] * b[i];
-                na += a[i] * a[i];
-                nb += b[i] * b[i];
-            }
-            if na == 0.0 || nb == 0.0 {
-                0.0
-            } else {
-                dot / (na.sqrt() * nb.sqrt())
-            }
-        }
+        // Use shared cosine similarity function
+        let cosine = |a: &[f32], b: &[f32]| -> f32 {
+            crate::utils::cosine_similarity(a, b)
+        };
 
         let mut matches: Vec<(f32, SimpleRow)> = Vec::new();
         let mut skipped_mismatched = 0;
