+++ src/utils/mod.rs
@@ -1,0 +1,6 @@
+//! Utility functions shared across the codebase
+
+pub mod math;
+
+// Re-export commonly used utilities
+pub use math::cosine_similarity;
